refactor: code refactor
1. Remove the redundant header files included.
2. Remove redundant code(file.h,file.cpp).
3. remove ll-store-installer

Log: Code optimization
