+++ libs/linglong/src/linglong/cli/cli.h
@@ -13,12 +13,6 @@
 
 #include <docopt.h>
 
-#include <QCommandLineOption>
-#include <QCommandLineParser>
-#include <QCoreApplication>
-
-#include <csignal>
-
 namespace linglong::cli {
 
 class Cli : public QObject
