+++ libs/linglong/src/linglong/package/architecture.cpp
@@ -11,7 +11,6 @@
 #include <cstring>
 #include <fstream>
 #include <iostream>
-#include <string>
 #include <vector>
 
 namespace linglong::package {
