+++ libs/linglong/src/linglong/runtime/container_builder.h
@@ -13,7 +13,6 @@
 #include "ocppi/runtime/config/types/Mount.hpp"
 
 #include <QDir>
-#include <QProcess>
 
 namespace linglong::runtime {
